requirements.txt
# Core LangChain packages
langchain>=1.2.0
langchain-core>=1.2.0
langchain-openai>=0.3.0
langchain-community>=0.3.0

# LangGraph (if using advanced workflows)
langgraph>=1.0.0

# Ollama integration
langchain-ollama>=0.1.0

# OpenAI integration
openai>=1.0.0

# Streamlit for web interface
streamlit>=1.31.0

# Environment variable management
python-dotenv>=1.0.0

# Core dependencies (auto-installed but listed for clarity)
pydantic>=2.7.4,<3.0.0
pyyaml>=5.3.0
tenacity>=8.1.0
requests>=2.0.0

# LangSmith tracking (optional - for debugging/monitoring)
langsmith>=0.3.45